refactored to better file structure with shared components
