+++ packages/frontend/src/modules/app/pages/EditTodoPage/EditTodoPage.style.ts
@@ -1,0 +1,8 @@
+import { css } from '@emotion/css';
+import { THEME } from '../../../../shared/styles/constants';
+
+export const checkboxContainerStyles = css`
+	display: flex;
+	justify-content: space-between;
+	margin-bottom: ${THEME.SPACING.MEDIUM};
+`;
