+++ packages/frontend/src/modules/app/pages/TodosPage/TodosPage.styles.ts
@@ -1,5 +1,4 @@
 import { css } from '@emotion/css';
-// import { THEME, COLORS } from '../../../shared/styles/constants';
 
 export const buttonGroup = css`
 	margin-bottom: 3rem;
